#496: split APIServerApp bootstrap into three units
`configure(_:)` was a 210-line function that mixed env loading, on-disk
directory setup, middleware order, view/static-file wiring, database
configuration, lifecycle handler registration, and SSO config warnings.
The MARK sections gave it some navigability, but anything cross-cutting
(e.g. "where does the worker secret get loaded?") still required a full
read.

Split into:

  Bootstrap/AppDirectories.swift   — on-disk dirs + worker secret +
                                     autostart + service stores
  Bootstrap/AppMiddleware.swift    — sessions + the order-sensitive
                                     middleware chain + Leaf tags +
                                     static-file middleware
  Bootstrap/AppServices.swift      — database + migrations +
                                     lifecycle handlers + BrightSpace
                                     + SSO config-validation warnings

`configure(_:)` now orchestrates the three. The preloaded-config /
auth-mode-override resolution stays in APIServerApp.swift as
`resolveAppConfig(...)` since it predates the bootstrap units. Both
production and the `Application.preloadedAppConfig` test seam continue
to work unchanged.

APIServerApp.swift goes 399 → 242 LOC. Bootstrap files are 51 + 114 +
69 = 234 LOC, with no functional duplication. Middleware ordering and
storage-seeding semantics are preserved exactly. Bootstrap-sensitive
test suites (RoleMiddleware, AuthModeGating, HealthRoutes) pass.
